requirements.txt
pandas
openai
requests
openpyxl
numpy
scikit-learn
nltk
llama-cpp-python
unidecode
